requirements.txt
# CloudAuditor Python Dependencies
# Python 3.14+ compatible versions

# AWS SDK for Python
boto3>=1.35.0

# PostgreSQL database adapter
psycopg2-binary>=2.9.9

# HTML/XML parsing library
beautifulsoup4>=4.12.0

# HTTP library for Python
requests>=2.32.0
